+++ debug/launch.ts
@@ -1,6 +1,3 @@
-#!/usr/bin/env node
-'use strict';
-
-const launcher = require('..');
+import * as launcher from '../src/launch';
 
 launcher.main(process.env.npm_lifecycle_event, process.argv, process.env.npm_config_argv);